Renames FederationStorageProvider derived and test classes. Moves getNewFederationBtcUTXOs tests from BridgeStorageProviderTest to FedreationStorageProviderImplTest
