+++ rskj-core/src/test/java/co/rsk/peg/BridgeStorageProviderTest.java
@@ -3164,46 +3164,6 @@ class BridgeStorageProviderTest {
     }
 
     @Test
-    void getNewFederationBtcUTXOs_before_RSKIP284_before_RSKIP293_testnet() throws IOException {
-        testGetNewFederationBtcUTXOs(false, false, NetworkParameters.ID_TESTNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_before_RSKIP284_before_RSKIP293_mainnet() throws IOException {
-        testGetNewFederationBtcUTXOs(false, false, NetworkParameters.ID_MAINNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_before_RSKIP284_after_RSKIP293_testnet() throws IOException {
-        testGetNewFederationBtcUTXOs(false, true, NetworkParameters.ID_TESTNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_before_RSKIP284_after_RSKIP293_mainnet() throws IOException {
-        testGetNewFederationBtcUTXOs(false, true, NetworkParameters.ID_MAINNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_after_RSKIP284_before_RSKIP293_testnet() throws IOException {
-        testGetNewFederationBtcUTXOs(true, false, NetworkParameters.ID_TESTNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_after_RSKIP284_before_RSKIP293_mainnet() throws IOException {
-        testGetNewFederationBtcUTXOs(true, false, NetworkParameters.ID_MAINNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_after_RSKIP284_after_RSKIP293_testnet() throws IOException {
-        testGetNewFederationBtcUTXOs(true, true, NetworkParameters.ID_TESTNET);
-    }
-
-    @Test
-    void getNewFederationBtcUTXOs_after_RSKIP284_after_RSKIP293_mainnet() throws IOException {
-        testGetNewFederationBtcUTXOs(true, true, NetworkParameters.ID_MAINNET);
-    }
-
-    @Test
     void saveNewFederationBtcUTXOs_no_data() throws IOException {
         ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
 
@@ -3432,68 +3392,6 @@ class BridgeStorageProviderTest {
             Assertions.assertEquals(2, storageBytesCalls.size());
             Assertions.assertEquals(1, serializeCalls.size());
         }
-    }
-
-    private void testGetNewFederationBtcUTXOs(boolean isRskip284Active, boolean isRskip293Active, String networkId) throws IOException {
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        when(activations.isActive(ConsensusRule.RSKIP284)).thenReturn(isRskip284Active);
-        when(activations.isActive(ConsensusRule.RSKIP293)).thenReturn(isRskip293Active);
-
-        NetworkParameters networkParameters = NetworkParameters.fromID(networkId);
-
-        Repository repository = mock(Repository.class);
-        List<UTXO> federationUtxos = Arrays.asList(
-            PegTestUtils.createUTXO(1, 0, Coin.COIN),
-            PegTestUtils.createUTXO(2, 2, Coin.COIN.divide(2)),
-            PegTestUtils.createUTXO(3, 0, Coin.COIN.multiply(3))
-        );
-        when(repository.getStorageBytes(
-            bridgeAddress,
-            NEW_FEDERATION_BTC_UTXOS_KEY.getKey()
-        )).thenReturn(BridgeSerializationUtils.serializeUTXOList(federationUtxos));
-
-        List<UTXO> federationUtxosAfterRskip284Activation = Arrays.asList(
-            PegTestUtils.createUTXO(4, 0, Coin.FIFTY_COINS),
-            PegTestUtils.createUTXO(5, 2, Coin.COIN.multiply(2))
-        );
-        when(repository.getStorageBytes(
-            bridgeAddress,
-            NEW_FEDERATION_BTC_UTXOS_KEY_FOR_TESTNET_PRE_HOP.getKey()
-        )).thenReturn(BridgeSerializationUtils.serializeUTXOList(federationUtxosAfterRskip284Activation));
-
-        List<UTXO> federationUtxosAfterRskip293Activation = Arrays.asList(
-            PegTestUtils.createUTXO(6, 1, Coin.valueOf(150_000)),
-            PegTestUtils.createUTXO(7, 3, Coin.COIN.multiply(3))
-        );
-        when(repository.getStorageBytes(
-            bridgeAddress,
-            NEW_FEDERATION_BTC_UTXOS_KEY_FOR_TESTNET_POST_HOP.getKey()
-        )).thenReturn(BridgeSerializationUtils.serializeUTXOList(federationUtxosAfterRskip293Activation));
-
-        FederationStorageProvider federationStorageProvider = createFederationStorageProvider(repository);
-
-        List<UTXO> obtainedUtxos = federationStorageProvider.getNewFederationBtcUTXOs(networkParameters, activations);
-
-        if (!networkId.equals(NetworkParameters.ID_TESTNET)) {
-            Assertions.assertEquals(federationUtxos, obtainedUtxos);
-            return;
-        }
-
-        // testnet
-        // rskip284 & rskip293 are not active
-        if (!isRskip284Active) {
-            Assertions.assertEquals(federationUtxos, obtainedUtxos);
-            return;
-        }
-
-        // rskip284 is active
-        if (!isRskip293Active) {
-            Assertions.assertEquals(federationUtxosAfterRskip284Activation, obtainedUtxos);
-            return;
-        }
-
-        // rskip293 is active
-        Assertions.assertEquals(federationUtxosAfterRskip293Activation, obtainedUtxos);
     }
 
     private void testSaveNewFederationBtcUTXOs(boolean isRskip284Active, String networkId) throws IOException {
