+++ rskj-core/src/test/java/co/rsk/peg/federation/FederationStorageProviderImplTests.java
@@ -11,26 +11,39 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import co.rsk.bitcoinj.core.Address;
 import co.rsk.bitcoinj.core.BtcECKey;
+import co.rsk.bitcoinj.core.Coin;
 import co.rsk.bitcoinj.core.NetworkParameters;
+import co.rsk.core.RskAddress;
 import co.rsk.peg.BridgeSerializationUtils;
+import co.rsk.peg.InMemoryStorage;
 import co.rsk.peg.PegTestUtils;
+import co.rsk.peg.bitcoin.BitcoinTestUtils;
 import co.rsk.peg.constants.BridgeConstants;
+import co.rsk.peg.constants.BridgeTestNetConstants;
+import java.io.IOException;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import co.rsk.peg.constants.BridgeRegTestConstants;
 import co.rsk.peg.federation.constants.FederationConstants;
 import co.rsk.peg.storage.BridgeStorageAccessorImpl;
 import co.rsk.peg.storage.StorageAccessor;
+import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ActivationConfigsForTest;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
 import org.ethereum.core.Repository;
+import org.ethereum.vm.DataWord;
 import org.ethereum.vm.PrecompiledContracts;
 import org.junit.jupiter.api.Test;
 import org.mockito.verification.VerificationMode;
+import co.rsk.bitcoinj.core.UTXO;
+import org.junit.jupiter.api.Assertions;
 
-class FederationStorageProviderTests {
+class FederationStorageProviderImplTests {
 
     private static final int STANDARD_MULTISIG_FEDERATION_FORMAT_VERSION = STANDARD_MULTISIG_FEDERATION.getFormatVersion();
     private static final int NON_STANDARD_ERP_FEDERATION_FORMAT_VERSION = NON_STANDARD_ERP_FEDERATION.getFormatVersion();
@@ -40,6 +53,10 @@ class FederationStorageProviderTests {
     private final FederationConstants federationConstantsRegtest = bridgeConstantsRegtest.getFederationConstants();
     private final NetworkParameters regtestBtcParams = bridgeConstantsRegtest.getBtcParams();
     private ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
+    private final BridgeConstants bridgeConstantsTestnet = BridgeTestNetConstants.getInstance();
+    private final FederationConstants federationConstantsTestnet = bridgeConstantsTestnet.getFederationConstants();
+    private final NetworkParameters testnetBtcParams = federationConstantsTestnet.getBtcParams();
+    private final RskAddress bridgeAddress = PrecompiledContracts.BRIDGE_ADDR;
 
     @Test
     void getNewFederation_should_return_p2sh_erp_federation() {
@@ -534,6 +551,145 @@ class FederationStorageProviderTests {
             OLD_FEDERATION_KEY.getKey(),
             null
         );
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_before_RSKIP284_before_RSKIP293_testnet() throws IOException {
+        testGetNewFederationBtcUTXOs(false, false, NetworkParameters.ID_TESTNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_before_RSKIP284_before_RSKIP293_mainnet() throws IOException {
+        testGetNewFederationBtcUTXOs(false, false, NetworkParameters.ID_MAINNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_before_RSKIP284_after_RSKIP293_testnet() throws IOException {
+        testGetNewFederationBtcUTXOs(false, true, NetworkParameters.ID_TESTNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_before_RSKIP284_after_RSKIP293_mainnet() throws IOException {
+        testGetNewFederationBtcUTXOs(false, true, NetworkParameters.ID_MAINNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_after_RSKIP284_before_RSKIP293_testnet() throws IOException {
+        testGetNewFederationBtcUTXOs(true, false, NetworkParameters.ID_TESTNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_after_RSKIP284_before_RSKIP293_mainnet() throws IOException {
+        testGetNewFederationBtcUTXOs(true, false, NetworkParameters.ID_MAINNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_after_RSKIP284_after_RSKIP293_testnet() throws IOException {
+        testGetNewFederationBtcUTXOs(true, true, NetworkParameters.ID_TESTNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_after_RSKIP284_after_RSKIP293_mainnet() throws IOException {
+        testGetNewFederationBtcUTXOs(true, true, NetworkParameters.ID_MAINNET);
+    }
+
+    @Test
+    void getNewFederationBtcUTXOs_calledTwice_returnCachedUtxos() {
+
+        ActivationConfig.ForBlock activations = ActivationConfigsForTest.only(ConsensusRule.RSKIP123).forBlock(0);
+        StorageAccessor storageAccessor = new InMemoryStorage();
+
+        DataWord newFederationBtcUtxosKey = NEW_FEDERATION_BTC_UTXOS_KEY.getKey();
+
+        Address btcAddress = new Address(testnetBtcParams, Hex.decode("4a22c3c4cbb31e4d03b15550636762bda0baf85a"));
+        List<UTXO> expectedUtxos = BitcoinTestUtils.createUTXOs(1, btcAddress);
+        storageAccessor.saveToRepository(newFederationBtcUtxosKey, expectedUtxos, BridgeSerializationUtils::serializeUTXOList);
+
+        FederationStorageProvider federationStorageProvider = new FederationStorageProviderImpl(storageAccessor);
+
+        List<UTXO> actualUtxos = federationStorageProvider.getNewFederationBtcUTXOs(testnetBtcParams, activations);
+
+        assertEquals(1, actualUtxos.size());
+        assertEquals(expectedUtxos, actualUtxos);
+
+        List<UTXO> extraUtxos = new ArrayList<>(expectedUtxos);
+        extraUtxos.addAll(BitcoinTestUtils.createUTXOs(1, btcAddress));
+
+        storageAccessor.saveToRepository(newFederationBtcUtxosKey, extraUtxos, BridgeSerializationUtils::serializeUTXOList);
+
+        List<UTXO> actualUtxosAfterSecondGet = federationStorageProvider.getNewFederationBtcUTXOs(null, activations);
+
+        assertEquals(1, actualUtxosAfterSecondGet.size());
+        assertEquals(expectedUtxos, actualUtxosAfterSecondGet);
+
+        List<UTXO> actualUtxosInStorage = storageAccessor.getFromRepository(newFederationBtcUtxosKey, BridgeSerializationUtils::deserializeUTXOList);
+
+        assertEquals(2, actualUtxosInStorage.size());
+        assertEquals(extraUtxos, actualUtxosInStorage);
+
+    }
+
+    private void testGetNewFederationBtcUTXOs(boolean isRskip284Active, boolean isRskip293Active, String networkId) throws IOException {
+        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
+        when(activations.isActive(ConsensusRule.RSKIP284)).thenReturn(isRskip284Active);
+        when(activations.isActive(ConsensusRule.RSKIP293)).thenReturn(isRskip293Active);
+
+        NetworkParameters networkParameters = NetworkParameters.fromID(networkId);
+
+        Repository repository = mock(Repository.class);
+        List<UTXO> federationUtxos = Arrays.asList(
+            PegTestUtils.createUTXO(1, 0, Coin.COIN),
+            PegTestUtils.createUTXO(2, 2, Coin.COIN.divide(2)),
+            PegTestUtils.createUTXO(3, 0, Coin.COIN.multiply(3))
+        );
+        when(repository.getStorageBytes(
+            bridgeAddress,
+            NEW_FEDERATION_BTC_UTXOS_KEY.getKey()
+        )).thenReturn(BridgeSerializationUtils.serializeUTXOList(federationUtxos));
+
+        List<UTXO> federationUtxosAfterRskip284Activation = Arrays.asList(
+            PegTestUtils.createUTXO(4, 0, Coin.FIFTY_COINS),
+            PegTestUtils.createUTXO(5, 2, Coin.COIN.multiply(2))
+        );
+        when(repository.getStorageBytes(
+            bridgeAddress,
+            NEW_FEDERATION_BTC_UTXOS_KEY_FOR_TESTNET_PRE_HOP.getKey()
+        )).thenReturn(BridgeSerializationUtils.serializeUTXOList(federationUtxosAfterRskip284Activation));
+
+        List<UTXO> federationUtxosAfterRskip293Activation = Arrays.asList(
+            PegTestUtils.createUTXO(6, 1, Coin.valueOf(150_000)),
+            PegTestUtils.createUTXO(7, 3, Coin.COIN.multiply(3))
+        );
+        when(repository.getStorageBytes(
+            bridgeAddress,
+            NEW_FEDERATION_BTC_UTXOS_KEY_FOR_TESTNET_POST_HOP.getKey()
+        )).thenReturn(BridgeSerializationUtils.serializeUTXOList(federationUtxosAfterRskip293Activation));
+
+        FederationStorageProvider federationStorageProvider = createFederationStorageProvider(repository);
+
+        List<UTXO> obtainedUtxos = federationStorageProvider.getNewFederationBtcUTXOs(networkParameters, activations);
+
+        if (!networkId.equals(NetworkParameters.ID_TESTNET)) {
+            Assertions.assertEquals(federationUtxos, obtainedUtxos);
+            return;
+        }
+
+        // testnet
+        // rskip284 & rskip293 are not active
+        if (!isRskip284Active) {
+            Assertions.assertEquals(federationUtxos, obtainedUtxos);
+            return;
+        }
+
+        // rskip284 is active
+        if (!isRskip293Active) {
+            Assertions.assertEquals(federationUtxosAfterRskip284Activation, obtainedUtxos);
+            return;
+        }
+
+        // rskip293 is active
+        Assertions.assertEquals(federationUtxosAfterRskip293Activation, obtainedUtxos);
+
     }
 
     private void testSaveOldFederation(
